Move AnimalCache to Data folder and namespace
